feat: cards_proj cleaned code
This code is a version from Python Testing with pytest, Second Edition book (https://pragprog.com/titles/bopytest2/python-testing-with-pytest-second-edition/). I set up this environment to use the pre-commit with hooks that cleaned the original code. For more details about the hooks, see the .pre-commit-config.yaml file in the repository root.